Add support for hashing pod annotations within Deployment
This allows for dependency hashing to work inside of a deployment pod
template. This is a common use case, where you want pods to be
re-created when a config map is updated. I _think_ this is the only way
to achieve this, as updating a deployment's annotations would not cause
Pods to be recreated.

As part of this, it's necessary to add the namespace parameter to the
hashDependency function, because this information isn't available in the
metadata that we extract from the PodSpec. This feels recursive, and we
could avoid the duplicated code by recursing on the Filter function, but
the issue there is that we lose the namespace when we recurse, which is
crucial for making sure we don't match resources across namespaces. We
could embed the namespace in the metadata and remove it again later, but
it felt simplest to just duplicate the code. We can come up with a more
elegant solution if we find more special cases.