Refactor hash dependency to use latest kyaml architecture
This refactors the hash dependency filter to make use of the latest
kyaml arch. The main change is that filters now accept their list of
inputs instead of a single node. This simplifies things because we don't
then need to store the list of nodes in the filter object for later use.

This also does some refactoring on the tests to factor out common utils.
Currently they don't pass, this will be fixed in the next commit.