Add the ability to specify an auth method to sync
This adds the ability to specify an auth method as an argument to the
sync function explicitly, instead of trying to determine the auth method
based on the arguments passed.

This enables us to also detect omission of an auth method, which is
perfectly fine if the repos you are fetching are public.