Update workflow to only publish one function at a time
This updates the Github actions workflow to only publish one function at
a time.

On master, we still publish all functions with the tag latest.

However if the tag begins with a v, we are expecting a tag of the format
`v1.2.3/function-name`. This tag is parsed, and then used to only
publish the tagged function.