Only publish one function at a time (#23)
* Update workflow to only publish one function at a time

This updates the Github actions workflow to only publish one function at
a time.

On master, we still publish all functions with the tag latest.

However if the tag begins with a v, we are expecting a tag of the format
`v1.2.3/function-name`. This tag is parsed, and then used to only
publish the tagged function.

* Add shell lint make targets, and fix issues

* Separate shell and golang lint, and add workflow step

* Fix workflow file syntax

* Replace local make target with action

* Fix argument name

* Add script to check whether current ref should be published

The logic of doing if statements in the Github workflows file is
limiting. Instead, we're going to use a quick shell script that encodes
this logic.

* Flip expected function position in git tags

This changes the expect tag syntax to be function/version rather than
version/function, in order to better integrate with Kpt.

* Integrate should publish script into github workflow